Add files via upload
Simple Keras NN model trained using hand landmarks. Its architecture is based on a model available on Kaggle.
At the moment the script for real time prediction uses this model so change it accordingly before submitting it. 
Additionally I will upload the notebook containing the code used to train it.